fix(evals): added the types
